Catch python exceptions by line
Adds new logic for catching errors that may be caused by incomplete or invalid
objects that appeared earlier in the file. Implements a few on the spot checks,
but the major check is a try-except at the line level that will throw an
uncategorized message if a syntax error occurs.